Puse para que se reinicie la bala al llegar al final del mapa, y puse dano a la bala y vida al bart. Ahora el bart se muere de dos balazos ATR
Atte chipi

+++ torreDefensa/src/Disparo/DisparoEnemigo.java
@@ -4,7 +4,7 @@ import Mapa.Mapa;
 
 public abstract class DisparoEnemigo extends Disparo  {
 
-	public DisparoEnemigo(Mapa miMapa,Celda miCelda, float daño, int velocidad) {
+	public DisparoEnemigo(Mapa miMapa,Celda miCelda, int daño, int velocidad) {
 		super(miMapa,miCelda, daño, velocidad);
 	}
